Use helper text field for date field.
We're using a text field to cause the keyboard to be displayed when the
field becomes the first responder. We will place the picker view in the
keyboard view. Since an insertion point does not make sense for a date
field, an opaque label is placed over the text field's text rect, which
hides the insertion point and prevents the user from using the loop.

This feels pretty hacky but this seems to be similar to the behavior
Safari has for select menus.